requirements.txt
sqlalchemy==2.0.23
pandas==2.1.3
pydantic==2.5.0
pyyaml==6.0.1
structlog==23.2.0
python-dotenv==1.0.0
jinja2==3.1.2
plotly==5.18.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pymssql==2.2.10
redis==5.0.1
elasticsearch==8.11.0
pymongo==4.6.1
cassandra-driver==3.29.0